+++ src/server/routes/profile/practice.js
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+const router = Router()
+
+router.get('/', (req, res) => {
+    res.render('practice.html')
+})
+
+export default router
